requirements.txt
django>=4.0.2
djangorestframework>=3.13.1
python-dotenv>=0.19.2
faker>=13.2.0
django-extensions>=3.1.5
coverage>=6.3.2
django-cors-headers>=3.11.0
channels >=3.0.4